backend/requirements.txt
flask
flask-cors
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pillow
intel-openmp
mkl
